pyproject.toml
[build-system]
  requires = ["poetry-core"]
  build-backend = "poetry.core.masonry.api"

[tool]
  [tool.poetry]
    package-mode = false
    [tool.poetry.dependencies]
      python = ">=3.9,<4.0.0"
      pulumi = "~3.122.0"
      pulumi-github = "~6.2.2"
      Jinja2 = "~3.1.4"
      PyYAML = "~6.0.1"

    [tool.poetry.group.dev.dependencies]
      ruff = "~0.5.0"